requirements.txt
discord.py==1.7.3
python-dotenv==0.19.2
requests==2.26.0
black==21.12b0
discord-components==2.1.2